requirements.txt
replicate==0.25.2
snowflake-snowpark-python==1.16.0
streamlit==1.34.0
python-dotenv==1.0.1